Lamp and stuff
Implemented the lamp with animation on pickup/change and drop.

Fixed a bunch of bugs to make sure it runs smoothly.

Going to try to see if I can replace all the platforms with real art,
not sure if I'll make it in time so I'll duplicate the LEVEL scene and
try to make it. So we have a backup if I fail!